==> notebooks/7_Location data preparation.ipynb <==
library(here)
library(xlsx)
library(tmap)
library(sf)
library(rgdal)
library(raster)
library(tidyr)
library(stringr)

### Load utility functions and themes
source(here("R","getNicheOutline.R"))
source(here("R","theme_tmaps.R"))

# Define imputs paths:

# Admnistrative map of China
path2china<-here('raw_data','CHN_adm')

# Archaeological records location data
path2buckwheat<-here("data","ESM_2.csv") # Buckwheat's occurrence records
path2locations<-here("data","guedesbocinsky2018_crops_across_eurasia_translated.xlsx")

# Production and predictions data
path2predictions <- here("outputs","predictions")

# Define output paths:

path2outputs<-here("manuscript","Chapter6","Figs","Raster")
path2sites<-here("data","site_locations.csv")
path2limits<-here("outputs","chronology_limits.csv")

# Import data:
# Import the outer border of China
outer_border<-st_read(dsn=path2china,layer='CHN_adm0')
# Import the spatial polygons data frame of counties:
china<-readOGR(dsn=path2china,layer='CHN_adm3')
# Import production data

### Load crop locations

locations<-read.xlsx(path2locations,1)

# Inculde only rows with data, that were not explicitly excluded
locations<-locations[!is.na(locations$Site)&is.na(locations$Exclude.),]

locations[,c("Longitude")]<-as.numeric(locations$Longitude)
locations[,c("Latitude")]<-as.numeric(locations$Latitude)

locations$lower_range<-locations$Age.range.lower..BP
locations$upper_range<-locations$Age.range.upper..BP

locations[is.na(locations$Age.range.lower..BP),"lower_range"] <- locations[is.na(locations$Age.range.lower..BP),]$X14C.age.BP + (2*locations[is.na(locations$Age.range.lower..BP),]$X1.sigma.uncertainty)
locations[is.na(locations$Age.range.lower..BP),"upper_range"] <- locations[is.na(locations$Age.range.upper..BP),]$X14C.age.BP - (2*locations[is.na(locations$Age.range.lower..BP),]$X1.sigma.uncertainty)

locations_df<-locations[,c("Site","Longitude","Latitude","X14C.date.on.cereal.","lower_range","upper_range","Wheat","Barley","Foxtail.millet","Broomcorn.millet","Millet..unidentified.","Rice","Rice..wild.")] # Get only required columns

locations_df<-locations_df[!(is.na(locations_df$lower_range))&!(is.na(locations_df$upper_range)),]

buckwheat<-read.csv(path2buckwheat)

buckwheat$Buckwheat<-ifelse(buckwheat$Taxonomic_identification=="Fagopyrum tataricum",NA,TRUE)
buckwheat$TartaryBuckwheat<-ifelse(buckwheat$Taxonomic_identification=="Fagopyrum tataricum",TRUE,NA)
buckwheat$X14C.date.on.cereal.<-ifelse(buckwheat$Sample_type=="Charred seeds"&buckwheat$Dating_method=="AMS 14C",TRUE,FALSE)

sample_type<-buckwheat[,c("Sample_type","Site_short","Buckwheat")]
colnames(sample_type)[2]<-"Site"

buckwheat<-buckwheat[,c("Site_short","longitude","latitude","X14C.date.on.cereal.","start_date","end_date","Buckwheat","TartaryBuckwheat")]
colnames(buckwheat)<-c("Site","Longitude","Latitude","X14C.date.on.cereal.","lower_range","upper_range","Buckwheat","TartaryBuckwheat")

common_sites<- buckwheat[buckwheat$Site %in% locations_df$Site,]$Site
cs_df<-unique(locations_df[locations_df$Site %in% common_sites,c("Site","Longitude","Latitude")])

buckwheat<-merge(buckwheat,cs_df, by=c("Site"),all=TRUE)

buckwheat$Longitude.x<- ifelse(is.na(buckwheat$Longitude.y),buckwheat$Longitude.x,buckwheat$Longitude.y)
buckwheat$Latitude.x<- ifelse(is.na(buckwheat$Latitude.y),buckwheat$Latitude.x,buckwheat$Latitude.y)

buckwheat<-buckwheat[,!colnames(buckwheat)%in%c("Longitude.y","Latitude.y")]

colnames(buckwheat)<-c("Site","Longitude","Latitude","X14C.date.on.cereal.","lower_range","upper_range","Buckwheat","TartaryBuckwheat")

locations_df<-merge(locations_df,buckwheat, by=c("Site","Longitude","Latitude","X14C.date.on.cereal.","lower_range","upper_range"),all=TRUE)

# Transform location data data into a spatial points data.frame
coordinates(locations_df)<-~Longitude+Latitude
crs(locations_df) <-crs(china)

locations_china<-locations_df[china,]

uq<-locations_china[,"Site"]
uq<-uq[which(!duplicated(uq$Site)),]

zd <- zerodist(uq)

locations<-as.data.frame(locations_china)

# Assign occurrence records to time slices (up till 8000 BP)

start=1500
loc<-subset(locations,upper_range<=1500)
loc$period <- "1000 BP"
for (i in c(1:17)){
    sub<-subset(locations,(upper_range<=start+1000 & upper_range>start)| (lower_range>start & lower_range<=start+1000 ))
    start=start+1000
    if(nrow(sub)>0){
       sub$period <- paste(start-500,"BP",sep=" ")
       loc<-rbind(loc,sub)}
}
#older_records<-subset(locations,upper_range>10500 | lower_range>10500)
#older_records$period <- "> 10 000 BP"
#loc<-rbind(loc,older_records)

### Aggregate by period

loc<-loc[,!names(loc)%in%c("lower_range","upper_range","X14C.date.on.cereal.")]

loc<-unique(loc)

# Aggregate sites assigned to the same period

loc<-aggregate(loc[,!(names(loc)%in%c("Site","Longitude","Latitude","period"))], 
             by = list(Site=loc$Site,Longitude=loc$Longitude,Latitude=loc$Latitude,period=loc$period),
             FUN=function(x) {if ((TRUE %in% x==TRUE)) return(TRUE) else return(NA)})

loc$period<-factor(loc$period,levels=paste(c(seq(1000,14000,1000),17000),"BP"),ordered=TRUE)

loc<-loc[!(loc$period>"10000 BP"&!is.na(loc$Buckwheat)),]

loc$Millet<-ifelse((!is.na(loc$Foxtail.millet)|!is.na(loc$Broomcorn.millet)|!is.na(loc$Millet..unidentified.)),TRUE,NA)

 loc<-merge(loc,sample_type,by=c("Site","Buckwheat"),all.x=TRUE)

loc[loc$Site=="Kyung-lung Mesa",]$Buckwheat<-NA

write.csv(loc,path2sites)

#crops<-c("rice","millet","buckwheat","wheat","barley")
#locByCrop<-gather(loc, crop,present,str_to_title(crops),na.rm=TRUE)
#d<-aggregate(period~crop,data=locByCrop,FUN="max")
